requirements.txt
feedparser==6.0.11
openai>=1.57.0
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1
